Initial Commit
I created a fairly simple main.rs that allows to call into different challenge solutions.
Just create a Struct which implements the Challenge trait and add it to the vec
at the start of the main() using boxed() like boxed(Day01::new()).
See the next commit of how to implement a challenge.