Implement Day 1
It took me a while to solve part two.
The initial RegEx I had did not correctly match lines like "35zrgthreetwonesz",
which should result in "3" and "one" as match, but resulted in "3" and "two".

Using the AoC solution of @eidonia from GitHub, I found that RegEx look-ahead was the solution.
Then after fiddeling around with Rust::regex, I replaced it with pcre2 which supports look-ahead.

After that, it gave me the same result as the implementation of eidonia which was correct.